pyproject.toml
[tool.poetry]
name = "event-management-with-swaggerr"
version = "0.1.0"
description = ""
authors = ["yes"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
annotated-types = "0.7.0"
asgiref = "3.8.1"
authlib = "1.3.2"
build = "1.2.2.post1"
cachecontrol = "0.14.1"
certifi = "2024.8.30"
cffi = "1.17.1"
cfgv = "3.4.0"
charset-normalizer = "3.4.0"
cleo = "2.1.0"
click = "8.1.7"
colorama = "0.4.6"
crashtest = "0.4.1"
cryptography = "44.0.0"
distlib = "0.3.9"
django = "5.1.4"
django-cors-headers = "4.6.0"
django-extensions = "3.2.3"
djangorestframework = "3.15.2"
djangorestframework-jwt = "1.11.0"
djangorestframework-simplejwt = "5.3.0"
dparse = "0.6.4"
drf-yasg = "1.21.8"
dulwich = "0.21.7"
fastjsonschema = "2.21.1"
filelock = "3.16.1"
identify = "2.6.3"
idna = "3.10"
importlib-metadata = "8.5.0"
inflection = "0.5.1"
installer = "0.7.0"
jaraco-classes = "3.4.0"
jinja2 = "3.1.4"
keyring = "24.3.1"
liccheck = "0.9.2"
markdown-it-py = "3.0.0"
markupsafe = "3.0.2"
marshmallow = "3.23.1"
mdurl = "0.1.2"
more-itertools = "10.5.0"
msgpack = "1.1.0"
nodeenv = "1.9.1"
packaging = "24.2"
pexpect = "4.9.0"
pkginfo = "1.12.0"
platformdirs = "4.3.6"
poetry-core = "1.9.1"
poetry-plugin-export = "1.8.0"
pre-commit = "4.0.1"
psutil = "6.0.0"
ptyprocess = "0.7.0"
pycparser = "2.22"
pydantic = "2.9.2"
pydantic-core = "2.23.4"
pygments = "2.18.0"
pyjwt = "1.7.1"
pyproject-hooks = "1.2.0"
python-dotenv = "1.0.1"
pytz = "2024.2"
pywin32-ctypes = "0.2.3"
pyyaml = "6.0.2"
rapidfuzz = "3.10.1"
requests = "2.32.3"
requests-toolbelt = "1.0.0"
rich = "13.9.4"
ruamel-yaml = "0.18.6"
ruamel-yaml-clib = "0.2.12"
safety-schemas = "0.0.10"
semantic-version = "2.10.0"
shellingham = "1.5.4"
sqlparse = "0.5.3"
toml = "0.10.2"
tomli = "2.2.1"
tomlkit = "0.13.2"
trove-classifiers = "2024.10.21.16"
typer = "0.15.1"
typing-extensions = "4.12.2"
tzdata = "2024.2"
uritemplate = "4.1.1"
urllib3 = "2.2.3"
virtualenv = "20.28.0"
zipp = "3.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
